requirements.txt
django
djangorestframework
django-filter
markdown
stripe
django-crispy-forms
